modularize commands module into extensions, use pseudo DB from snakecore, delete emotion-related functionality and unused modules, implement bootstrap.json, various refactorings
