+++ bootstrap.json
@@ -1,0 +1,13 @@
+{
+    "extensions": [
+        {
+            "name": "pgbot.commands.help"
+        },
+        {
+            "name": "pgbot.commands.admin"
+        },
+        {
+            "name": "pgbot.commands.user"
+        }
+    ]
+}
