+++ pgbot/commands/utils/help.py
@@ -72,11 +72,10 @@ def get_doc_from_func(func: typing.Callable):
 
 async def send_help_message(
     ctx: commands.Context,
+    bot: commands.Bot,
     original_msg: discord.Message,
     invoker: discord.Member,
-    commands: tuple[str, ...],
-    cmds_and_funcs: dict[str, typing.Callable],
-    groups: dict[str, list],
+    qualified_name: typing.Optional[str] = None,
     page: int = 1,
 ):
     """
@@ -87,24 +86,23 @@ async def send_help_message(
     Args:
         original_msg: The message to edit
         invoker: The member who requested the help command
-        commands: A tuple of command names passed by user for help.
-        cmds_and_funcs: The name-function pairs to get the docstrings from
-        groups: The name-list pairs of group commands
         page: The page of the embed, 0 by default
     """
 
     doc_fields = {}
     embeds = []
 
-    if not commands:
-        functions = {}
-        for key, func in cmds_and_funcs.items():
-            if hasattr(func, "groupname"):
-                key = f"{func.groupname} {' '.join(func.subcmds)}"
-            functions[key] = func
+    is_admin = any(
+        role.id in common.ServerConstants.ADMIN_ROLES
+        for role in getattr(invoker, "roles", ())
+    )
 
-        for func in functions.values():
-            data = get_doc_from_func(func)
+    if not qualified_name:
+        for cmd in sorted(bot.walk_commands(), key=lambda cmd: cmd.qualified_name):
+            if cmd.hidden or not is_admin and cmd.extras.get("admin_only", False):
+                continue
+
+            data = get_doc_from_func(cmd.callback)
             if not data:
                 continue
 
@@ -142,75 +140,82 @@ async def send_help_message(
                 )
             )
 
-    elif commands[0] in cmds_and_funcs:
-        func_name = commands[0]
-        funcs = groups[func_name] if func_name in groups else []
-        funcs.insert(0, cmds_and_funcs[func_name])
-
-        for func in funcs:
-            if (
-                commands[1:]
-                and commands[1:] != getattr(func, "subcmds", ())[: len(commands[1:])]
-            ):
-                continue
-
-            doc = get_doc_from_func(func)
-            if not doc:
-                # function found, but does not have help.
-                return await snakecore.utils.embed_utils.replace_embed_at(
-                    original_msg,
-                    title="Could not get docs",
-                    description="Command has no documentation",
-                    color=0xFF0000,
-                )
-
-            body = f"`{doc['signature']}`\n`Category: {doc['type']}`\n\n"
-
-            desc = doc["description"]
-
-            ext_desc = doc.get("extended description")
-            if ext_desc:
-                desc = f"> *{desc}*\n\n{ext_desc}"
-
-            desc_list = desc.split(sep="+===+")
-
-            body += f"**Description:**\n{desc_list[0]}"
-
-            embed_fields = []
-
-            example_cmd = doc.get("example command")
-            if example_cmd:
-                embed_fields.append(
-                    dict(name="Example command(s):", value=example_cmd, inline=True)
-                )
-
-            if len(desc_list) == 1:
-                embeds.append(
-                    snakecore.utils.embed_utils.create_embed(
-                        title=f"Help for `{func_name}`",
-                        description=body,
-                        color=common.BOT_HELP_PROMPT["color"],
-                        fields=embed_fields,
+    else:
+        cmd = bot.get_command(qualified_name)
+        if (
+            cmd is not None
+            and not cmd.hidden
+            and (is_admin or cmd.extras.get("admin_only", False))
+        ):
+            cmds = [cmd]
+            if isinstance(cmd, commands.Group):
+                cmds.extend(
+                    sorted(
+                        (subcmd for subcmd in cmd.walk_commands()),
+                        key=lambda cmd: cmd.qualified_name,
                     )
                 )
-            else:
-                embeds.append(
-                    snakecore.utils.embed_utils.create_embed(
-                        title=f"Help for `{func_name}`",
-                        description=body,
-                        color=common.BOT_HELP_PROMPT["color"],
+
+            for cmd in cmds:
+                doc = get_doc_from_func(cmd.callback)
+                if not doc:
+                    # function found, but does not have help.
+                    return await snakecore.utils.embed_utils.replace_embed_at(
+                        original_msg,
+                        title="Could not get docs",
+                        description="Command has no documentation",
+                        color=0xFF0000,
                     )
-                )
-                desc_list_len = len(desc_list)
-                for i in range(1, desc_list_len):
+
+                body = f"`{doc['signature']}`\n`Category: {doc['type']}`\n\n"
+
+                desc = doc["description"]
+
+                ext_desc = doc.get("extended description")
+                if ext_desc:
+                    desc = f"> *{desc}*\n\n{ext_desc}"
+
+                desc_list = desc.split(sep="+===+")
+
+                body += f"**Description:**\n{desc_list[0]}"
+
+                embed_fields = []
+
+                example_cmd = doc.get("example command")
+                if example_cmd:
+                    embed_fields.append(
+                        dict(name="Example command(s):", value=example_cmd, inline=True)
+                    )
+
+                cmd_qualified_name = cmd.qualified_name
+
+                if len(desc_list) == 1:
                     embeds.append(
                         snakecore.utils.embed_utils.create_embed(
-                            title=f"Help for `{func_name}`",
-                            description=desc_list[i],
+                            title=f"Help for `{cmd_qualified_name}`",
+                            description=body,
                             color=common.BOT_HELP_PROMPT["color"],
-                            fields=embed_fields if i == desc_list_len - 1 else None,
+                            fields=embed_fields,
                         )
                     )
+                else:
+                    embeds.append(
+                        snakecore.utils.embed_utils.create_embed(
+                            title=f"Help for `{cmd_qualified_name}`",
+                            description=body,
+                            color=common.BOT_HELP_PROMPT["color"],
+                        )
+                    )
+                    desc_list_len = len(desc_list)
+                    for i in range(1, desc_list_len):
+                        embeds.append(
+                            snakecore.utils.embed_utils.create_embed(
+                                title=f"Help for `{cmd_qualified_name}`",
+                                description=desc_list[i],
+                                color=common.BOT_HELP_PROMPT["color"],
+                                fields=embed_fields if i == desc_list_len - 1 else None,
+                            )
+                        )
 
     if not embeds:
         return await snakecore.utils.embed_utils.replace_embed_at(
