+++ pgbot/routine.py
@@ -9,25 +9,22 @@ It gets called every 5 seconds or so.
 
 import asyncio
 import datetime
-from http.client import HTTPException
 import io
 import os
-import random
 import sys
-from typing import Union
 
 import discord
 from discord.ext import tasks
 import snakecore
 
-from pgbot import common, db, emotion
+from pgbot import common
 
 
-async def handle_reminders(reminder_obj: db.DiscordDB):
+async def handle_reminders(reminder_obj: snakecore.db.DiscordDB):
     """
     Handle reminder routines
     """
-    reminders = reminder_obj.get({})
+    reminders = reminder_obj.obj
 
     new_reminders = {}
     for mem_id, reminder_dict in reminders.items():
@@ -117,11 +114,8 @@ async def routine():
     Function that gets called routinely. This function inturn, calles other
     routine functions to handle stuff
     """
-    async with db.DiscordDB("reminders") as db_obj:
+    async with snakecore.db.DiscordDB("reminders") as db_obj:
         await handle_reminders(db_obj)
-
-    if random.randint(0, 4) == 0:
-        await emotion.update("bored", 1)
 
     await common.bot.change_presence(
         activity=discord.Activity(
